+++ src/models/message.py
@@ -1,0 +1,5 @@
+class Message():
+    def __init__(self, user: str, text: str, message_type: str):
+        self.user = user
+        self.text = text
+        self.message_type = message_type
